requirements.txt
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
numpy==1.24.2
pandas==2.0.0
pangres==4.1.4
psycopg2-binary==2.9.6
SQLAlchemy==2.0.9
